refactor: simplify panic setup
