+++ Cargo.toml
@@ -29,9 +29,6 @@ dtb-parser = { path = "libs/dtb-parser" }
 kmm = { path = "libs/kmm" }
 ktest = { path = "libs/ktest" }
 linked-list-allocator = { path = "libs/linked-list-allocator" }
-panic-abort = { path = "libs/panic-abort" }
-panic-common = { path = "libs/panic-common" }
-panic-unwind = { path = "libs/panic-unwind" }
 riscv = { path = "libs/riscv" }
 semihosting-logger = { path = "libs/semihosting-logger" }
 sync = { path = "libs/sync" }
