+++ libs/ktest/Cargo.toml
@@ -16,7 +16,6 @@ ktest-macros = { path = "macros" }
 kernel.workspace = true
 loader-api.workspace = true
 dtb-parser.workspace = true
-panic-unwind.workspace = true
 
 [target.'cfg(any(target_arch = "riscv64", target_arch = "riscv32"))'.dependencies]
 riscv.workspace = true
