--- libs/panic-common/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "panic-common"
-version.workspace = true
-edition.workspace = true
-authors.workspace = true
-license.workspace = true
-
-[lints]
-workspace = true
-
-[dependencies]
-sync.workspace = true
-log.workspace = true
-cfg-if.workspace = true
-
-[target.'cfg(any(target_arch = "riscv64", target_arch = "riscv32"))'.dependencies]
-riscv.workspace = true
-
