--- libs/panic-unwind/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "panic-unwind"
-version.workspace = true
-edition.workspace = true
-authors.workspace = true
-license.workspace = true
-
-[lib]
-harness = false
-
-[lints]
-workspace = true
-
-[dependencies]
-panic-common.workspace = true
-thread-local.workspace = true
-log.workspace = true
-unwind2.workspace = true
