+++ loader/Cargo.toml
@@ -17,18 +17,17 @@ lz4_flex.workspace = true
 
 [dependencies]
 loader-api.workspace = true
-kmm.workspace = true
-panic-abort.workspace = true
-semihosting-logger = { workspace = true }
 sync.workspace = true
-linked-list-allocator.workspace = true
 dtb-parser.workspace = true
+semihosting-logger = { workspace = true }
+kmm.workspace = true
 
 cfg-if.workspace = true
 log.workspace = true
 arrayvec.workspace = true
 object = { workspace = true, features = ["read_core", "elf"] }
 lz4_flex.workspace = true
+linked-list-allocator.workspace = true
 onlyerror.workspace = true
 rand_chacha = { version = "0.3.1", default-features = false }
 rand = { version = "0.8.5", default-features = false }
